authentication/requirements.txt
alembic==1.8.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
backports.entry-points-selectable==1.2.0
bcrypt==4.0.1
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
execnet==2.0.2
fastapi==0.103.1
filelock==3.12.4
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
identify==2.5.29
idna==3.4
iniconfig==2.0.0
isort==5.12.0
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathlib==1.0.1
pathspec==0.11.2
platformdirs==2.6.2
pluggy==1.3.0
pre-commit==3.4.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.0
pydantic-settings==2.0.3
pydantic_core==2.10.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-slugify==8.0.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.0b3
starlette==0.27.0
text-unidecode==1.3
trio==0.22.2
typing_extensions==4.8.0
uvicorn==0.23.2
virtualenv==20.10.0
psycopg2-binary==2.9.7
pandas==2.1.1
sqlalchemy-utils==0.41.1